+++ packages/dashboard/src/features/ai/pages/AIQuickStartPage.tsx
@@ -3,8 +3,14 @@ import { Button, CopyButton, Tab, Tabs } from '@insforge/ui';
 import { CodeEditor } from '#components';
 import { useOpenRouterKey } from '#features/ai/hooks/useOpenRouterKey';
 import { cn } from '#lib/utils/utils';
-
-type QuickStartMode = 'text' | 'image' | 'video';
+import {
+  PROMPT_CARD_COPY,
+  QUICK_START_COPY,
+  QUICK_START_MODES,
+  getQuickStartPrompt,
+  getQuickStartScript,
+  type QuickStartMode,
+} from '#features/ai/constants';
 
 interface QuickStartStep {
   id: number;
@@ -22,139 +28,6 @@ interface CodeBlockProps {
   copyText?: string;
   badge?: string;
   kind: 'shell' | 'env' | 'javascript';
-}
-
-const QUICK_START_MODES: { value: QuickStartMode; label: string }[] = [
-  { value: 'text', label: 'Text Generation' },
-  { value: 'image', label: 'Image Generation' },
-  { value: 'video', label: 'Video Generation' },
-];
-
-const PROMPT_CARD_COPY: Record<QuickStartMode, string> = {
-  text: 'Copy this prompt for your agent to generate text through the OpenRouter model gateway.',
-  image: 'Copy this prompt for your agent to generate images through the OpenRouter model gateway.',
-  video: 'Copy this prompt for your agent to generate videos through the OpenRouter model gateway.',
-};
-
-const MODE_COPY: Record<
-  QuickStartMode,
-  { projectName: string; description: string; model: string; installCommand: string }
-> = {
-  text: {
-    projectName: 'ai-text-demo',
-    description: 'Create a chat completion through OpenRouter with the OpenAI SDK.',
-    model: 'openai/gpt-5.5',
-    installCommand: 'npm install openai dotenv\nnpm install --save-dev @types/node tsx typescript',
-  },
-  image: {
-    projectName: 'ai-image-demo',
-    description: 'Generate an image with an OpenRouter model that supports image output.',
-    model: 'google/gemini-2.5-flash-image',
-    installCommand: 'npm install openai dotenv\nnpm install --save-dev @types/node tsx typescript',
-  },
-  video: {
-    projectName: 'ai-video-demo',
-    description: 'Submit an asynchronous video generation job and poll until it completes.',
-    model: 'google/veo-3.1',
-    installCommand: 'npm install dotenv\nnpm install --save-dev @types/node tsx typescript',
-  },
-};
-
-function getScript(mode: QuickStartMode, model: string) {
-  if (mode === 'image') {
-    return `import OpenAI from 'openai';
-import 'dotenv/config';
-
-const openai = new OpenAI({
-  baseURL: 'https://openrouter.ai/api/v1',
-  apiKey: process.env.OPENROUTER_API_KEY,
-});
-
-const completion = await openai.chat.completions.create({
-  model: '${model}',
-  modalities: ['image', 'text'],
-  messages: [
-    { role: 'user', content: 'Generate a beautiful sunset over mountains.' },
-  ],
-});
-
-const message = completion.choices[0]?.message;
-console.log(message?.content);
-console.log(message?.images?.[0]?.image_url?.url);`;
-  }
-
-  if (mode === 'video') {
-    return `import 'dotenv/config';
-
-const response = await fetch('https://openrouter.ai/api/v1/videos', {
-  method: 'POST',
-  headers: {
-    Authorization: \`Bearer \${process.env.OPENROUTER_API_KEY}\`,
-    'Content-Type': 'application/json',
-  },
-  body: JSON.stringify({
-    model: '${model}',
-    prompt: 'A golden retriever playing fetch on a sunny beach.',
-  }),
-});
-
-const job = await response.json();
-console.log('Video job:', job.id);
-
-let result = job;
-while (result.status !== 'completed' && result.status !== 'failed') {
-  await new Promise((resolve) => setTimeout(resolve, 5000));
-  const poll = await fetch(\`https://openrouter.ai/api/v1/videos/\${job.id}\`, {
-    headers: { Authorization: \`Bearer \${process.env.OPENROUTER_API_KEY}\` },
-  });
-  result = await poll.json();
-  console.log('Status:', result.status);
-}
-
-console.log(result);`;
-  }
-
-  return `import OpenAI from 'openai';
-import 'dotenv/config';
-
-const openai = new OpenAI({
-  baseURL: 'https://openrouter.ai/api/v1',
-  apiKey: process.env.OPENROUTER_API_KEY,
-});
-
-const completion = await openai.chat.completions.create({
-  model: '${model}',
-  messages: [
-    { role: 'user', content: 'What is the meaning of life?' },
-  ],
-});
-
-console.log(completion.choices[0]?.message?.content);`;
-}
-
-function getQuickStartPrompt(mode: QuickStartMode) {
-  const copy = MODE_COPY[mode];
-  const featureCopy: Record<QuickStartMode, string> = {
-    text: 'a text generation feature that sends a user prompt and renders the model response',
-    image: 'an image generation feature that sends a user prompt and renders the returned image',
-    video:
-      'a video generation feature that submits a prompt, polls the job status, and renders the completed video',
-  };
-  const apiCopy: Record<QuickStartMode, string> = {
-    text: 'Use the OpenAI SDK with baseURL set to https://openrouter.ai/api/v1.',
-    image:
-      "Use the OpenAI SDK with baseURL set to https://openrouter.ai/api/v1 and request modalities ['image', 'text'].",
-    video:
-      'Use fetch with the OpenRouter video endpoint at https://openrouter.ai/api/v1/videos; do not install or use the OpenAI SDK for video.',
-  };
-
-  return [
-    `Add ${featureCopy[mode]} using the OpenRouter model gateway.`,
-    `Use model ${copy.model}. ${apiCopy[mode]}`,
-    'First inspect the existing project and integrate with its current framework, routing, styling, and state patterns. If it is React, Next.js, Vue, Svelte, or another framework, add the feature inside that app instead of creating a separate demo project.',
-    'Store the API key in OPENROUTER_API_KEY and read it from environment variables. Do not hard-code secrets or expose server-only keys to the browser; add a backend/API route when the framework needs one.',
-    'Install only the dependencies needed for this project, keep the UI minimal and consistent with the existing design, handle loading and error states, and include brief run instructions after implementation.',
-  ].join('\n');
 }
 
 function ShellLine({ line }: { line: string }) {
@@ -353,7 +226,7 @@ function StepItem({ step, isLast }: { step: QuickStartStep; isLast: boolean }) {
 export default function AIQuickStartPage() {
   const [mode, setMode] = useState<QuickStartMode>('text');
   const { data: openRouterKey, isLoading: isOpenRouterKeyLoading } = useOpenRouterKey();
-  const copy = MODE_COPY[mode];
+  const copy = QUICK_START_COPY[mode];
   const quickStartPrompt = useMemo(() => getQuickStartPrompt(mode), [mode]);
   const displayedOpenRouterKey = isOpenRouterKeyLoading
     ? 'Loading...'
@@ -379,7 +252,7 @@ export default function AIQuickStartPage() {
     {
       id: 2,
       title: 'Install Dependencies',
-      description: 'Install the OpenAI SDK and development dependencies.',
+      description: copy.description,
       blocks: [
         {
           code: copy.installCommand,
@@ -412,7 +285,7 @@ export default function AIQuickStartPage() {
       blocks: [
         {
           badge: 'index.ts',
-          code: getScript(mode, copy.model),
+          code: getQuickStartScript(mode, copy.model),
           kind: 'javascript',
         },
       ],
